third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/xmlhttprequest-overridemimetype-content-type-header-expected.txt

This tests that XMLHttpRequest overrideMimeType() does not update the Content-Type header for the response.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr.getResponseHeader("Content-Type") is "application/xml"
PASS successfullyParsed is true

TEST COMPLETE
